# add TextMatchFilterOptimizer to maximally push down `text_match` filters to Lucene