**Motivation:** 
Query performance against the Lucene index suffers when chaining multiple `text_match` predicates together. Our users often programmatically generate their queries, which exacerbates the issue as 10s/100s of `text_match` predicates can be included in a single query. 

Because of this, users are required to understand Pinot's Lucene implementation details for them to compose an efficient query. To remove this requirement, this PR adds a `TextMatchFilterOptmizer` that performs the optimization automatically. 

**Summary:**
This functionality is best understood through the unit testcases. In short: 
- Merge all AND's and OR's `text_match` operands when possible, without affecting query accuracy
- Push down NOT into Lucene, unless all `text_match` filters are inversed, then the NOT expression remains in Pinot

**Open question:**
There is one edge case (that I can think of) where this optimization can hurt performance: if there are a number of `text_match OR text_match OR text_match` etc, early termination when `limit` is reached might take longer since the entire merged `text_match` query must now complete. For this reason, it might be prudent to put this behind a query option (or rather a query option to disable it, since I believe it makes more sense to enable by default). 

Ideally, the `LuceneDocIdCollector` could early terminate (but doesn't currently have the required context). 

Testing: unit tests (query performance separately verified via running the optimized vs unoptimized queries). Sample optimization improvements:
```
count(*) = 500_000_000
optimized: 2950ms
unoptmized: 8350ms
> select count(*) from table where text_match(column, '"message_logtype:storage-event" AND /.*c28487d062.*/')
> select count(*) from table where text_match(column, '"message_logtype:storage-event"') AND text_match(column, '/.*c28487d062.*/')


count(*) = 2
optimized: 12ms
unoptimized: 8400ms
> select count(*) from table where text_match(column, '"message_logtype:storage-event" AND /.*c28487d062.*/ AND "offset:46127612"')
> select count(*) from table where text_match(column, '"message_logtype:storage-event"') AND text_match(column, '/.*c28487d062.*/') AND text_match(column, '"offset:46127612"')
```

tags: `feature`, `performance` (?)