@deemoliu 2024-03-15
this looks awesome. is it possible to optimize regexp_like(x) OR regexp_like(y) ?